Implement 'run' and 'check' commands
The 'debops run' and 'debops check' commands can be used to execute
Ansible playbooks against hosts, or check any changes when they are
applied, respectively. Playbooks can be specified either as direct paths
to files with or without the '.yml' extansion, or defined as
'<namespace>.<collection/><path/to/playbook>' strings which will be
expanded by DebOps to use playbooks included in Ansible Collections.