Don't nest project configuration too deep
Previously in the DebOps configuration structure, the current project
configuration was nested under ['projects'][name] key. This could be
used to modify specific project options from the global configuration,
however this caused the per-project configuration to require nesting
each time.

This patch changes the code so that the per-project configuration is on
the same level as global configuration. This should make the per-project
config files easier to write and understand. The global state between
DebOps project can be handled some other way in the future.